+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add core/butttons + core/button to ConstructBlock
 * Refactor ConstructBlock class for better handling
 * Add ConstructACfBlock helper
+* Add possibility to register only the selected script or style.
 
 ## 0.5.2 (2023-04-25)